Created CommonFilterInputs and CommonFilterMultipleInputs
