+++ layers/Schema/packages/comments/src/ModuleProcessors/CommentFilterInnerModuleProcessor.php
@@ -7,6 +7,7 @@ namespace PoPSchema\Comments\ModuleProcessors;
 use PoP\Hooks\Facades\HooksAPIFacade;
 use PoP\ComponentModel\ModuleProcessors\AbstractModuleProcessor;
 use PoPSchema\SchemaCommons\ModuleProcessors\FormInputs\CommonFilterInputs;
+use PoPSchema\SchemaCommons\ModuleProcessors\FormInputs\CommonFilterMultipleInputs;
 
 class CommentFilterInnerModuleProcessor extends AbstractModuleProcessor
 {
@@ -29,7 +30,7 @@ class CommentFilterInnerModuleProcessor extends AbstractModuleProcessor
                 [CommonFilterInputs::class, CommonFilterInputs::MODULE_FILTERINPUT_ORDER],
                 [CommonFilterInputs::class, CommonFilterInputs::MODULE_FILTERINPUT_LIMIT],
                 [CommonFilterInputs::class, CommonFilterInputs::MODULE_FILTERINPUT_OFFSET],
-                [CommonFilterInputs::class, CommonFilterInputs::MODULE_FILTERINPUT_DATES],
+                [CommonFilterMultipleInputs::class, CommonFilterMultipleInputs::MODULE_FILTERINPUT_DATES],
                 [CommonFilterInputs::class, CommonFilterInputs::MODULE_FILTERINPUT_IDS],
                 [CommonFilterInputs::class, CommonFilterInputs::MODULE_FILTERINPUT_ID],
             ],
