+++ layers/Schema/packages/events/src/ModuleProcessors/FilterInnerModuleProcessor.php
@@ -6,6 +6,7 @@ namespace PoPSchema\Events\ModuleProcessors;
 
 use PoP\ComponentModel\ModuleProcessors\AbstractModuleProcessor;
 use PoPSchema\SchemaCommons\ModuleProcessors\FormInputs\CommonFilterInputs;
+use PoPSchema\SchemaCommons\ModuleProcessors\FormInputs\CommonFilterMultipleInputs;
 
 class FilterInnerModuleProcessor extends AbstractModuleProcessor
 {
@@ -30,13 +31,13 @@ class FilterInnerModuleProcessor extends AbstractModuleProcessor
                 [CommonFilterInputs::class, CommonFilterInputs::MODULE_FILTERINPUT_ORDER],
                 [CommonFilterInputs::class, CommonFilterInputs::MODULE_FILTERINPUT_LIMIT],
                 [CommonFilterInputs::class, CommonFilterInputs::MODULE_FILTERINPUT_OFFSET],
-                [CommonFilterInputs::class, CommonFilterInputs::MODULE_FILTERINPUT_DATES],
+                [CommonFilterMultipleInputs::class, CommonFilterMultipleInputs::MODULE_FILTERINPUT_DATES],
                 [CommonFilterInputs::class, CommonFilterInputs::MODULE_FILTERINPUT_IDS],
                 [CommonFilterInputs::class, CommonFilterInputs::MODULE_FILTERINPUT_ID],
             ],
             self::MODULE_FILTERINNER_EVENTCOUNT => [
                 [CommonFilterInputs::class, CommonFilterInputs::MODULE_FILTERINPUT_SEARCH],
-                [CommonFilterInputs::class, CommonFilterInputs::MODULE_FILTERINPUT_DATES],
+                [CommonFilterMultipleInputs::class, CommonFilterMultipleInputs::MODULE_FILTERINPUT_DATES],
                 [CommonFilterInputs::class, CommonFilterInputs::MODULE_FILTERINPUT_IDS],
                 [CommonFilterInputs::class, CommonFilterInputs::MODULE_FILTERINPUT_ID],
             ],
